machintstCore核心的增加, 自定义脚本的优化, spline宕机问题的处理, icc瓦尔琪的处理

+++ src/server/game/Entities/Player/Machinist_VIP_Core.h
@@ -3,14 +3,14 @@
 
 struct VipElements
 {
-	uint32 vip;
+	uint8 vip;
 	uint32 jf;
 };
 
 struct ItemVIP
 {
 	uint32 item_id;
-	uint32 vip;
+	uint8 vip;
 };
 
 
@@ -32,16 +32,16 @@ public:
 	~VIP();
 
 	// Getterz
-	static uint32 GetVIPMAX();
+	static uint8 GetVIPMAX();
 	static float GetVIPOFFSET();
-	static uint32 GetTALENTBONUS();
+	static uint8 GetTALENTBONUS();
 	static float GetVIPLootMoneyBonus();
-	static uint32 GetVIP(uint32 acct_id);
+	static uint8 GetVIP(uint32 acct_id);
 	static uint32 GetPlayerJf(uint32 acct_id);
-	static uint32 GetItemVIP(uint32 item_id);
+	static uint8 GetItemVIP(uint32 item_id);
 
 	// Setterz
-	static void SetVIP(uint32 acct_id, uint32 pvip);
+	static void SetVIP(uint32 acct_id, uint8 pvip);
 	static void SetHearthStone(uint32 guid, uint32 map_id, float x, float y, float z, float o);
 	static void SetPlayerJf(uint32 acct_id, uint32 jf);
 	static bool VipHearthstone(Player* player);
